[tf][tfg] Add motivating tests for removing CastOp (and its folder)
Adds test cases for why the CastOp folder was removed (and its ODS definition that was marked as NoSideEffect).

The op cannot be marked as side-effect free (in MLIR terms) because the greedy pattern rewriter will remove ops from `tfg.graph` that are "trivially dead", although any op in the graph could have a use as a fetch. Also, ensure that casts even between identical types are not folded, because these dummy casts may be used to contain metadata

PiperOrigin-RevId: 429362019
Change-Id: I4d07bb1e16b5021fef28d6d11b1e9b7111202b7d