Remove cuda_py_tests and gpu_py_tests.
Users can switch to the direct expansion of cuda_py_tests into cuda_py_test. It is a bit laboring but these are BUILD files anyway.

The impact for removal of gpu_py_tests is likely very small. A search on github showed 250 results. I scanned a few pages (~10) and they are all forks of tensorflow org projects.

PiperOrigin-RevId: 430790702