[tf][tfg] Add a dialect interface to the TF op registry
Adds a dialect interface to the TFG dialect for querying the TF op registry.
Passes should be querying the op registry through this interface instead of
directly querying the global registry. The default implementation of the registry interface returns conservative values for op properties. A concrete implementation using an op registry must be injected by the user. This is to avoid having the TFG dialect depend on the TF framework library.

PiperOrigin-RevId: 432281406