Stages data on GPU host pinned memory during H2D transfer, if the tensor data is allocated on pageable memory.
This helps to preserve the concurrent kernel execution and data transfer, which could be disabled if GPU driver does the staging work.

The staging in this change does not really take effect, because memory types of all allocators have not been specified (i.e., kUnknown). The followup changes will specify memory types for allocators.

D2H transfer can be done similarly if it's needed.

PiperOrigin-RevId: 433582924