Improve debugging message when a cycle is detected
With this change, we'll print the names of all the nodes in the cycle. Without this node, we'll only print their numeric ids. Given only the numeric ids (and not the names), I have no idea how to actually track down the nodes responsible for creating a cycle in the graph.

PiperOrigin-RevId: 435419568