[tf][tfg] Introduce RegionPreservedAttrs
When a functional op is converted to a region-based op, it is eventually converted back to functional form for export. However, the original function cannot be used, as the region body as well as the function argument order may be modified. However, it important that when the new function is created, attributes from the previous function are inherited. E.g. whether the function should be JIT compiled.

RegionPreservedAttr is the mechanism to preserve function attributes when a functional control-flow op is converted to region form. Each region-based op has one (optional) such attribute for each of its regions. The attribute contains preserved function attributes, function argument attributes, and function result attributes.

If the region-based op is modified such that the number of arguments or results change, then the preserved attributes, if present, must be kept up-to-date.

PiperOrigin-RevId: 435752026