Quantize TF ops that should have the same input and output scales
For data movement ops, the input and output scales are the same and the computation can be done in int8, rather than dequantizing it to f32 to save memory bandwidth and improve performance.

This won't apply for the same scales ops that are not data movement ops - e.g. AveragePool.

PiperOrigin-RevId: 436149101