[XLA] Enable HloEvaluator fastpath in HloConstantFolding.
This lets us use Eigen for matmuls, for example.

Constant folding is often one of the slower passes in real-world models.

I also deleted the VLOG from the beginning/end of HloConstantFolding that dumps
the whole module.  This makes vlog so spammy as to be useless when dealing with
a large module.  If you want HLO before/after a pass, you can use
XLA_FLAGS=--xla_dump_hlo_pass_re=constant_folding.
PiperOrigin-RevId: 437302466