Fix training accuracy on AArch64 in FC layers
This patch fixes training accuracy on AArch64. When training on AArch64 using Compute Library as backend in oneDNN we need to inform Compute Library’s fully connected layer that weights are not constant so that they can be updated. This information now is propagated from MklDnnMatMulFwdPrimitive as a new parameter.