[tf][tfg] Create conversion pass from functional to region CF ops
Create the `tfg-functional-to-region` pass. This pass converts functional
control-flow operations to region-based operations: `tfg.If`, `tfg.Case`,
`tfg.While`, and `tfg.For`. The branch operations, If and Case, are created
with all sucessor operands converted to implicit capture. However, the loop
operations, While and For, keep all loop body arguments as explicit captures.

PiperOrigin-RevId: 438714296