[tf][tfg] Create conversion from region to functional control flow
Creates the pass `tfg-region-to-functional` which converts region-based control
flow ops to functional. The pass is executed in two steps: all implicitly
captured values used in the regions are converted to successor operands. Then,
new functions are created to house these regions, where they are moved.

The pass has an option `force-control-capture` to handle the following case:

```mlir
%ctl = NoOp
%If, %ctl_0 = If %cond then {
  %A, %ctl_1 = A [%Ctl]
} else { ... }
```

The `NoOp` control output cannot be converted to implicit capture without an
associated data edge. Conversion of this region op to functional form will fail.
If `force-control-capture` is set, then a `Const` is inserted to convert the
control edge to a data edge.

PiperOrigin-RevId: 440397807