Introduce constant folding pass in TFG
Create a direct translation of Grappler constant folding pass on TFG dialect. The pass is intended to be a drop in replacement/bug compatible and will be evolved post.

This CL implements all the transform patterns in grappler's constant folding but it hasn't reached the best status. To get the optimal result we still need a type inference pass which will be added later soon.

PiperOrigin-RevId: 441036078