[tf.data] Update docs for Dataset.cache().
Previously it wasn't clear what we meant by "iterated through in its entirety". This could be understood to mean calling `next` N times on a dataset with cardinality N. That is sufficient to memory-cache datasets with known cardinality, but for file-cached datasets or datasets with unknown cardinality, we don't finalize the cache until the input iterator to `cache` produces end_of_input.

PiperOrigin-RevId: 442909679