[XLA] Remove tuple-select instruction
It's already not supported on GPU, not used anywhere, does not exist in MHLO,
and adds a non-trivial amount of complexity to a large number of passes, since
it violates an invariant that a buffer for each instruction is known at compile
time.

For the XlaBuilder API, creates a conditional instead. To the best of our
knowledge, this API is not used anywhere (XlaBuilder::Select actually creates a
tuple-select if an input is a tuple).

For simplicity, this CL does not remove all the complexity arising from
tuple-select, some of it can be removed later.

PiperOrigin-RevId: 443060532