[tf][tfg] Region-Control-Flow: Set an instantiation limit on regions
MLIR has the expectation that regions aren't nested too deeply, which allows
many region-based algorithms be written recursively -- most importantly, the
verifier. If conditional regions get too deep, the verifier stack overflows.

This CL sets a limit on the number of iterations in FunctionalToRegion, which
sets an approximate instantiation depth. It also changes the driver to use
top-down traversal, which should improve compile times. And it skips regionDCE,
since all regions are single-block.

PiperOrigin-RevId: 444632406