Use rolling release for Bazel.
Since TF is in the progress of migrating to `cc_shared_library`, we are stress testing Bazel and uncovering a large number of bugs. We are in the state where some TF builds are broken with cascading failures and there is a long wait until a new LTS Bazel release. So, we're switching to rolling releases in an attempt to fix the builds.

However, if this causes more failures than it resolves, we will need to return back to stable releases.

PiperOrigin-RevId: 445259819