Add `PjRtDevice::Attributes()`, use it to implement `device.{name}` in Python.
There exist pybind11 types (`GpuDevice`, `TpuDevice`) just to allow device
specific attributes (e.g. `GpuDevice.device_vendor`, `TpuDevice.coords`) to be
queried by users. This change moves the ability to declare device specific
attributes to `PjRtDevice::Attributes()` and adds `Device.__getattr__` to
support accessing these properties.

In a future change we should be able to remove `GpuDevice` and `TpuDevice`
entirely since they are now redundant. I am not doing so in this change incase a
user is relying on these types in a hard to detect way and we get a rollback.

The main advantage of this approach is that `PjRtDevice` implementors can expose
custom device specific attributes to users without needing to maintain a
pybind11 caster for their type.

We may want to consider putting device specific attributes into a sub-property
of `Device` (e.g. `device.attributes.{name}`) to avoid any clashes with new
methods/properties we want to add in the future, however for now to maintain
backwards compatibility I have implemented `Device.__getattr__`.

PiperOrigin-RevId: 446292842