Remove [SameOperandsAndResultType] from mhlo.add
This paves the way towards supporting operands and results of different quantized types, as part of the ongoing work on sparsity in MHLO.

SameOperandsAndResultType has special interactions with type inference, so in order to keep parity with the current behavior, this CL also introduces [InferTensorType] and implements type inference logic accordingly.

The verification aspect of SameOperandsAndResultType, to an appropriate degree, will be reintroduced in a child CL in order to not muddy the waters. Type inference changes already lead to test failures.

PiperOrigin-RevId: 446622860