[TF2XLA] Light outside compilation for GPU
This CL adds a `CallTfKernel` generic TF2XLA kernel, which creates a custom
call that can call a TF kernel directly from HLO runtime.

The idea is that using this approach it is incredibly easy to add registrations
for TF kernels which don't have a good HLO translation.

In this design, the details of invoking the TF kernel are encoded in the JSON
format inside the custom-call HLO directly.

The advantages of such an approach include less extra copies, and the
replayability of the HLO dump (including the TF kernels).

The disadvantage is that all kernels still have to be registered for TF2XLA
translation (but the registration is trivial). In addition, light outside
compilation does not work for resource variables.

PiperOrigin-RevId: 447451767